success

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: lgernido <lgernido@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/19 13:12:07 by lgernido          #+#    #+#             */
-/*   Updated: 2023/11/21 15:53:49 by lgernido         ###   ########.fr       */
+/*   Updated: 2023/11/22 11:28:19 by lgernido         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,13 +25,13 @@ int		ft_printui(unsigned int n);
 int		ft_hexasize(unsigned int nb);
 int		ft_printhex(unsigned int nb, const char format);
 int		ft_format(va_list ap, const char format);
-int		ft_ptrsize(unsigned long long nb);
-int		ft_printptr(unsigned long long nb, const char format);
+int		ft_ptrsize(unsigned long nb);
+int		ft_printptr(unsigned long nb, const char format);
 int		ft_printf(const char *format, ...);
 
 char	*ft_uitoa(unsigned int n);
 
-void	ft_writeptr(unsigned long long nb, const char format);
+void	ft_writeptr(unsigned long nb, const char format);
 void	ft_writehex(unsigned int nb, const char format);
 
 #endif
